fix(badge): update stories for badge component
