--- core/components/atoms/badge/_stories_/SolidAlert.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Badge from '../Badge';
-
-export const solidAlert = () => (
-  <Badge appearance="alert" subtle={false}>
-    {'Failed'}
-  </Badge>
-);
-
-export default {
-  title: 'Components/Badge/Solid Alert',
-  component: Badge,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Badge',
-      },
-    },
-  },
-};
