--- core/components/atoms/badge/_stories_/SolidSecondary.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Badge from '../Badge';
-
-export const solidSecondary = () => (
-  <Badge appearance="secondary" subtle={false}>
-    {'Closed'}
-  </Badge>
-);
-
-export default {
-  title: 'Components/Badge',
-  component: Badge,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Badge',
-      },
-    },
-  },
-};
