--- core/components/atoms/badge/_stories_/SubtleAlert.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Badge from '../Badge';
-
-export const subtleAlert = () => (
-  <Badge appearance="alert" subtle={true}>
-    {'1.41'}
-  </Badge>
-);
-
-export default {
-  title: 'Components/Badge/Subtle Alert',
-  component: Badge,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Badge',
-      },
-    },
-  },
-};
