--- core/components/atoms/badge/_stories_/SubtleSecondary.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Badge from '../Badge';
-
-export const subtleSecondary = () => (
-  <Badge appearance="secondary" subtle={true}>
-    {'Closed'}
-  </Badge>
-);
-
-export default {
-  title: 'Components/Badge/Subtle Secondary',
-  component: Badge,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Badge',
-      },
-    },
-  },
-};
