+++ docs/src/pages/components/badges/code.mdx
@@ -1,7 +1,7 @@
 #### Live Demo
 <br/>
 
-<Preview name='components-badge--solid-secondary' />
+<Preview name='components-badge-all--all' />
 
 #### Props
-<PreviewWithPropTable name='components-badge--solid-secondary' />
+<PreviewWithPropTable name='components-badge-all--all' />
