+++ docs/src/pages/components/badges/usage.mdx
@@ -6,21 +6,27 @@ showMobile: true
 keywords: ['Tag', 'Label', 'Lozenge', 'Flag', 'Stamp']
 ---
 
-Badge contains short labels that help in 'tagging' entities. The container of a badge is rounded at the corners.
-<br/>
+Badges help the users to understand the difference between multiple entities by displaying concise metadata of that entity as a label.
+
+### Variants
+
+#### Solid badge
+
+Solid style is used to highlight important information on the page so that they can be recognized easily. Solid style helps the badge to easily pop out in a page.
+
 <Preview name='components-badge-solid-warning--solid-warning' />
 
-### Appearances
-A badge comes in **9 appearances** and **2 styles** each.
+#### Subtle badge
 
-#### Solid style
-Solid style is used to highlight important information on the page so that they can be recognized easily. Solid style helps the badge to easily pop out in a page.
-<Preview name='components-badge-variants-appearance--appearance' />
-<br/>
+Subtle style is used when a lot of entities on a single page need tagging/labelling. In that case, using solid badges becomes a bit overwhelming e.g. tags in a list. This is why subtle badges use the lightest shade of the colors.
 
-#### Subtle style
-Subtle style is used when a lot of entities on a single page need tagging/labeling. In that case, using solid badges becomes a bit overwhelming e.g. tags in a list. This is why subtle badges use the lightest shade of the colors.
 <Preview name='components-badge-subtle-warning--subtle-warning' />
+
+### Appearance 
+
+A badge comes in 9 appearances -
+
+<Preview name='components-badge-solid--solid' /> 
 
 ### Structure
 <br/>
